HUD, code cleanup (removed unneeded usings), fixed more doors (one left), added all weapon pickup. Removed old door system and made the new one much more extendable without needing to copy too much code.

+++ MetroidClone/MetroidClone/MetroidClone/Engine/InputHelper.cs
@@ -20,10 +20,8 @@ namespace MetroidClone.Engine
         private Stopwatch vibrateStopwatch = new Stopwatch();
         private double vibrateTime;
 
-        /// <summary>
-        /// updates the keyboard and mouse states or the gamepad state. can be switched between by pressing Enter on keyboard or
-        /// Start on gamepad. If there is no controller connected, it will automatically switch back to keyboard controls.
-        /// </summary>
+        //Updates the keyboard and mouse states or the gamepad state. can be switched between by pressing Enter on keyboard or
+        //Start on gamepad. If there is no controller connected, it will automatically switch back to keyboard controls.
         public void Update()
         {
             if (!ControllerInUse)
@@ -56,10 +54,12 @@ namespace MetroidClone.Engine
         {
             return keyBoardState.IsKeyDown(k);
         }
+
         public bool KeyboardCheckReleased(Keys k)
         {
             return keyBoardState.IsKeyUp(k) && lastKeyboardState.IsKeyDown(k);
         }
+
         public bool KeyboardCheckPressed(Keys k)
         {
             return keyBoardState.IsKeyDown(k) && lastKeyboardState.IsKeyUp(k);
@@ -75,7 +75,7 @@ namespace MetroidClone.Engine
         public Point MouseCheckUnscaledPosition(DrawWrapper drawWrapper)
         {
             Rectangle displayRect = drawWrapper.DisplayRect;
-            return new Point((mouseState.X - displayRect.X) * ((int) World.TileWidth * WorldGenerator.LevelWidth) / displayRect.Width, (mouseState.Y - displayRect.Y) * ((int) World.TileHeight * WorldGenerator.LevelHeight) / displayRect.Height);
+            return new Point((mouseState.X - displayRect.X) * World.TileWidth * WorldGenerator.LevelWidth / displayRect.Width, (mouseState.Y - displayRect.Y) * World.TileHeight * WorldGenerator.LevelHeight / displayRect.Height);
         }
 
         public bool MouseButtonCheckDown(bool left)
@@ -84,24 +84,28 @@ namespace MetroidClone.Engine
                 return mouseState.LeftButton == ButtonState.Pressed;
             return mouseState.RightButton == ButtonState.Pressed;
         }
+
         public bool MouseButtonCheckReleased(bool left)
         {
             if (left)
                 return mouseState.LeftButton == ButtonState.Released;
             return mouseState.RightButton == ButtonState.Released;
         }
+
         public bool MouseButtonCheckPressed(bool left)
         {
             if (left)
                 return mouseState.LeftButton == ButtonState.Pressed && lastMouseState.LeftButton == ButtonState.Released;
             return mouseState.RightButton == ButtonState.Pressed && lastMouseState.RightButton == ButtonState.Released;
         }
+
         public bool MouseWheelCheckScroll(bool up)
         {
             if (up)
                 return mouseState.ScrollWheelValue > lastMouseState.ScrollWheelValue;
             return mouseState.ScrollWheelValue < lastMouseState.ScrollWheelValue;
         }
+
         public bool MouseWheelPressed()
         {
             return mouseState.MiddleButton == ButtonState.Pressed && lastMouseState.MiddleButton == ButtonState.Released;
@@ -111,6 +115,7 @@ namespace MetroidClone.Engine
         {
             return gamePadState.IsConnected;
         }
+
         public Vector2 ThumbStickCheckDirection(bool left)
         {
             if (!ControllerInUse)
@@ -120,6 +125,7 @@ namespace MetroidClone.Engine
                 return gamePadState.ThumbSticks.Left;
             return gamePadState.ThumbSticks.Right;
         }
+        
         public bool ThumbStickCheckDown(bool left)
         {
             if (!ControllerInUse)
@@ -129,6 +135,7 @@ namespace MetroidClone.Engine
                 return (gamePadState.ThumbSticks.Left.X != 0 || gamePadState.ThumbSticks.Left.Y != 0);
             return (gamePadState.ThumbSticks.Right.X != 0 || gamePadState.ThumbSticks.Right.Y != 0);
         }
+
         public bool GamePadCheckDown(Buttons b)
         {
             if (!ControllerInUse)
@@ -136,6 +143,7 @@ namespace MetroidClone.Engine
 
             return gamePadState.IsButtonDown(b);
         }
+
         public bool GamePadCheckReleased(Buttons b)
         {
             if (!ControllerInUse)
@@ -143,6 +151,7 @@ namespace MetroidClone.Engine
 
             return gamePadState.IsButtonUp(b) && lastGamePadState.IsButtonDown(b);
         }
+
         public bool GamePadCheckPressed(Buttons b)
         {
             if (!ControllerInUse)
@@ -150,24 +159,31 @@ namespace MetroidClone.Engine
 
             return gamePadState.IsButtonDown(b) && lastGamePadState.IsButtonUp(b);
         }
+
         public bool GamePadTriggerCheckDown(bool left)
         {
             if (left)
                 return gamePadState.Triggers.Left != 0;
+
             return gamePadState.Triggers.Right != 0;
         }
+
         public bool GamePadTriggerCheckPressed(bool left)
         {
             if (left)
                 return gamePadState.Triggers.Left != 0 && lastGamePadState.Triggers.Left == 0;
+
             return gamePadState.Triggers.Right != 0 && lastGamePadState.Triggers.Right == 0;
         }
+
         public bool GamePadTriggerCheckReleased(bool left)
         {
             if (left)
                 return gamePadState.Triggers.Left == 0 && lastGamePadState.Triggers.Left != 0;
+
             return gamePadState.Triggers.Right == 0 && lastGamePadState.Triggers.Right != 0;
         }
+
         public void GamePadVibrate(float leftMotor, float rightMotor, double time)
         {
             vibrateStopwatch.Restart();
